feat[scanner, group badge]: Added post-scan for anon, changed message when validation fails.
