+++ packages/scanner/src/components/post-scan_new/post-scan-anon/index.ts
@@ -1,0 +1,1 @@
+export * from './post-scan-anon';
